repos/SBS-Backend-main/python/draft-image-generator/requirements.txt
# Function dependencies, for example:
# package>=version
# Function dependencies, for example:
# package>=version
Pillow
google-cloud-storage
uuid
google-cloud-logging